move the feature flags to the public directory
